Add config option to ignore devices
This is useful in a large building scenario, when WMS uses a single network
for multiple tenants, and we don't want to mess with our neighbours' shades.

On _why_ somebody would use a single network for multiple apartments,
the Weather Station seems to be the culprit - it can only serve one network
at a time.